Restructuring of file input/output
I've added an aggregation function, and changed the way the files are saved coming out of xDeepDive.

We now save files by "resource" so it's easier to manually check the results and apply regex or other search filters.

I've added the journal title to the CSV files, in case we want to begin to search that way.

There is now an aggregation function that allows us to aggregate resources by DOI to help build the bipartite network.